requirements.txt
autopep8==1.4.4
certifi==2019.9.11
Click==7.0
mysql-connector==2.2.9
numpy==1.17.3
pandas==0.25.3
psycopg2-binary==2.8.4
pycodestyle==2.5.0
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
SQLAlchemy==1.3.10
